+++ Extra/LocalizationService.swift
@@ -21,14 +21,34 @@ final class LocalizationService {
 
     // MARK: Initializers
     private init() {
-        for localeID in SupportedLocale.localeIDs where !SupportedLocale.systemLocaleIDs.contains(localeID) {
-            fatalError("Locale `\(localeID)` is not added to the project")
+        for localeID in SupportedLocale.allCases.map({ $0.id }) where !bundleLocaleIDs.contains(localeID) {
+            fatalError("Locale `\(localeID)` is not added to the project bundle")
         }
-        
-        for localeID in SupportedLocale.systemLocaleIDs where !SupportedLocale.localeIDs.contains(localeID) {
+
+        for localeID in bundleLocaleIDs where !SupportedLocale.allCases.map({ $0.id }).contains(localeID) {
             fatalError("Locale `\(localeID)` is not added to `LocalizationService.SupportedLocale`")
         }
     }
+    
+    // MARK: Configuration
+    func configure() {
+        _ = Self.shared
+    }
+    
+    // MARK: System and Bundle Locales
+    let preferredSystemLocaleIDs: [String] = Locale.preferredLanguages
+        .map { $0.removingRegionCode() }
+        .removingBaseLoacale()
+    
+    let preferredSystemLocaleID: String = Locale.current.identifier
+        .removingRegionCode()
+    
+    let bundleLocaleIDs: [String] = Bundle.main.localizations
+        .map { $0.removingRegionCode() }
+        .removingBaseLoacale()
+    
+    let bundleLocaleID: String? = Bundle.main.preferredLocalizations.first?
+        .removingRegionCode()
     
     // MARK: Supported Locale
     enum SupportedLocale: Identifiable, CaseIterable, KeyPathInitializableEnumeration {
@@ -36,7 +56,18 @@ final class LocalizationService {
         case english
         
         // MARK: Initializers
-        static var `default`: Self { .aCase(key: \.id, value: preferredSystemLocaleID)! } // fatalError
+        static var `default`: Self = {
+            if let preferredSystemLocale: Self = .aCase(key: \.id, value: LocalizationService.shared.preferredSystemLocaleID) {
+                return preferredSystemLocale
+            } else if
+                let bundleLocaleID = LocalizationService.shared.bundleLocaleID,
+                let bundleLocale: Self = .aCase(key: \.id, value: bundleLocaleID)
+            {
+                return bundleLocale
+            } else {
+                return .english
+            }
+        }()
         
         // MARK: Properties
         var id: String {
@@ -49,26 +80,13 @@ final class LocalizationService {
             NSLocale(localeIdentifier: LocalizationService.shared.locale.id)
                 .displayName(forKey: .identifier, value: id)
         }
-        
-        static let localeIDs: [String] = locales.map { $0.id }
-        static var locales: [Self] { Self.allCases }
-        
-        static let systemLocaleIDs: [String] = {
-            Locale.preferredLanguages
-                .map { $0.components(separatedBy: "-").first! } // fatalError
-                .filter { Bundle.main.localizations.filter { $0 != "Base" }.contains($0) }
-        }()
-        static let systemLocales: [Self] = systemLocaleIDs.map { .aCase(key: \.id, value: $0)! } // fatalError
-        
-        static let preferredSystemLocaleID: String = Locale.current.identifier.components(separatedBy: "_").first! // fatalError
-        static let preferredSystemLocale: Self = .aCase(key: \.id, value: preferredSystemLocaleID)! // fatalError
     }
     
     // MARK: Get / Set
     private func getLocale() -> SupportedLocale {
         guard
             let languagesIDs: [String] = UserDefaults.standard.value(forKey: Self.appleLanguagesUserDefaultsKey) as? [String],
-            let languageID: String = languagesIDs.first?.components(separatedBy: "-").first,
+            let languageID: String = languagesIDs.first?.removingRegionCode(),
             let locale: SupportedLocale = .aCase(key: \.id, value: languageID)
         else {
             setLocale(.default)
@@ -117,5 +135,18 @@ extension String {
             value: "",
             comment: ""
         )
+    }
+}
+
+// MARK: - Helpers
+extension String {
+    fileprivate func removingRegionCode() -> String {
+        components(separatedBy: .init(["-", "_"])).first ?? self
+    }
+}
+
+extension Array where Element == String {
+    fileprivate func removingBaseLoacale() -> [String] {
+        filter { $0 != "Base" }
     }
 }